Separate histogram tests
